Replaced the issue of arrow keys getting ignored with a delightful new issue, a bug discovery in iOS where the cursor position is actually kept secret.
If you always change the text in the field, then it ignores the second up/down until it parses a left/right.
If you always change the caret position, then it goes into a kind of infinite scrolling mode where it detects ups as downs until it comes back to the middle. (iOS bug).

Ideas - change the position BEFORE changing the text, or only change the text/position after up/down or left/right are detected?