Made the controller sprite a bit smaller. Allowed the controller to rotate during southpaw mode change Simplified highlighting in BPMControllerSprite Added MAW mode parsing
Tested with actual controllers.
Found that the keys were all lower case instead of upper.
Added logging of unrecognized keys to the dev tool.
Having some issues with the controllers:
The sliver one, C9C, doesn't seem to want to pair.
The Bronze one, C97, doesn't seem to want to change modes.
Noticed that the orientation of the controller in southpaw mode is backwards.